[TASK] Rector II
- RecastingRemovalRector

+++ Classes/Pagination/CustomAbstractPaginator.php
@@ -73,7 +73,7 @@ abstract class CustomAbstractPaginator extends AbstractPaginator
 
         $isUpdated = false;
         if ($this->currentPageNumber === $this->numberOfPages && $this->initialLimit > 0) {
-            $difference = $this->initialLimit - ((int)($this->itemsPerPage * ($this->currentPageNumber - 1)));
+            $difference = $this->initialLimit - ($this->itemsPerPage * ($this->currentPageNumber - 1));
             if ($difference > 0) {
                 $this->updatePaginatedItems($difference, $offset);
                 $isUpdated = true;
